Create Localization Manager
Add a module that allows to load named localization
files containing named strings for basic multi-language
support. If the locale string is not found in the file, the locale
itself is being returned.

This file also needs to be properly commented and a
documentation needs to be created for it.

The localization file that is added is also just an example
and needs the language strings added.