Create Autotext module
Add the event based autotext module which allows
to define an interval and returns an event every time
the interval runs out with a random message attached.

This file needs to be properly commented and needs
documentation.

The default_msgs.txt also has to be filled with the
required phrases for the contest.