Add config & json-fs-store modules from npm
This includes creating a ./config/ subdirectory and adding the default
./config/default.js config file. The default_msgs.txt has been removed
in favor of an entry in the config. The fileparser is no longer needed
since config supports multiple filetypes already.

The .gitignore has been updated to not include the json-fs-store
directory or the local* config override files in the ./config/ directory