Make locales JSON files
The locales are written in JSON format and should have the proper
file extension. The localization module now loads .json files instead
of .txt files.
Additionally, a de_DE.json locale has been added for the german
language.
The final change is that the localize function now accepts the config
object on creation to make it compatible with the configwrapper