Add special variables to replace
There is certain information that should be replacable in
localized strings. They will properly be replaced with
information coming from config now and can be used
with locales from the language file or just regular strings.

Also, "localization:" can be added to any string passed to
the function to make it more obvious that it comes from
the localization file. This is optional though, and does not
serve a purpose apart from making it clearer