Multiple minor changes to chat_connector
An argument "include_self" has been added that allows
to have the bot emit events if it is responsible for them itself.

Furthermore, a few more events are being directly forwarded
now, in case they ever need to be hooked.

A bit of debug code has been removed.

The userlist object and the user arrays have been moved so
that they are now accessible from the outside